Update fastfMRI.R
in jaccard.index, brought in an option for what value to assign if both x and y have no actives. By default this is set to 1, but the user is allowed to set this value now. For instance, if it is know that there should be some active elements, and both x and y report no actives, then the user could decide on a value (such as zero) to be assigned.